+++ webcomponents/color/src/components.d.ts
@@ -8,18 +8,28 @@ import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { DeckdeckgoPalette, DeckdeckgoPaletteColor } from "./utils/deckdeckgo-palette";
 export namespace Components {
     interface DeckgoColor {
+        /**
+          * The current selected color provided as hexadecimal value
+         */
         "colorHex": string;
+        /**
+          * The current selected color provided as a rgb value (without "rgb()", only value such as for example 255, 67, 54)
+         */
         "colorRgb": string;
-        "label": boolean;
-        "more": boolean;
-        "moreAlt": string;
+        /**
+          * An accessibility label for the color input field
+         */
+        "inputAlt": string;
+        /**
+          * The palette of color.
+         */
         "palette": DeckdeckgoPalette[];
     }
-    interface DeckgoColorLabel {
+    interface DeckgoColorInput {
         "colorHex": string;
-        "colorLabel": string;
         "colorRgb": string;
         "customColorRgb": string;
+        "inputAlt": string;
     }
 }
 declare global {
@@ -29,36 +39,50 @@ declare global {
         prototype: HTMLDeckgoColorElement;
         new (): HTMLDeckgoColorElement;
     };
-    interface HTMLDeckgoColorLabelElement extends Components.DeckgoColorLabel, HTMLStencilElement {
+    interface HTMLDeckgoColorInputElement extends Components.DeckgoColorInput, HTMLStencilElement {
     }
-    var HTMLDeckgoColorLabelElement: {
-        prototype: HTMLDeckgoColorLabelElement;
-        new (): HTMLDeckgoColorLabelElement;
+    var HTMLDeckgoColorInputElement: {
+        prototype: HTMLDeckgoColorInputElement;
+        new (): HTMLDeckgoColorInputElement;
     };
     interface HTMLElementTagNameMap {
         "deckgo-color": HTMLDeckgoColorElement;
-        "deckgo-color-label": HTMLDeckgoColorLabelElement;
+        "deckgo-color-input": HTMLDeckgoColorInputElement;
     }
 }
 declare namespace LocalJSX {
     interface DeckgoColor {
+        /**
+          * The current selected color provided as hexadecimal value
+         */
         "colorHex"?: string;
+        /**
+          * The current selected color provided as a rgb value (without "rgb()", only value such as for example 255, 67, 54)
+         */
         "colorRgb"?: string;
-        "label"?: boolean;
-        "more"?: boolean;
-        "moreAlt"?: string;
+        /**
+          * An accessibility label for the color input field
+         */
+        "inputAlt"?: string;
+        /**
+          * Emit the selected color
+         */
         "onColorChange"?: (event: CustomEvent<DeckdeckgoPaletteColor>) => void;
+        /**
+          * The palette of color.
+         */
         "palette"?: DeckdeckgoPalette[];
     }
-    interface DeckgoColorLabel {
+    interface DeckgoColorInput {
         "colorHex"?: string;
-        "colorLabel"?: string;
         "colorRgb"?: string;
         "customColorRgb"?: string;
+        "inputAlt"?: string;
+        "onSelectHexColor"?: (event: CustomEvent<DeckdeckgoPaletteColor>) => void;
     }
     interface IntrinsicElements {
         "deckgo-color": DeckgoColor;
-        "deckgo-color-label": DeckgoColorLabel;
+        "deckgo-color-input": DeckgoColorInput;
     }
 }
 export { LocalJSX as JSX };
@@ -66,7 +90,7 @@ declare module "@stencil/core" {
     export namespace JSX {
         interface IntrinsicElements {
             "deckgo-color": LocalJSX.DeckgoColor & JSXBase.HTMLAttributes<HTMLDeckgoColorElement>;
-            "deckgo-color-label": LocalJSX.DeckgoColorLabel & JSXBase.HTMLAttributes<HTMLDeckgoColorLabelElement>;
+            "deckgo-color-input": LocalJSX.DeckgoColorInput & JSXBase.HTMLAttributes<HTMLDeckgoColorInputElement>;
         }
     }
 }